% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{coerce,numeric,sparse_numeric-method}
\alias{coerce,numeric,sparse_numeric-method}
\title{Coercion from numeric to sparse_numeric}
\arguments{
\item{from}{A numeric vector to convert}
}
\value{
A sparse_numeric object
}
\description{
Coercion from numeric to sparse_numeric
}
