% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{coerce_sparse}
\alias{coerce_sparse}
\alias{coerce,numeric,sparse_numeric-method}
\alias{coerce,sparse_numeric,numeric-method}
\title{Coercion between numeric and sparse_numeric}
\arguments{
\item{from}{A \code{sparse_numeric} object.}
}
\value{
A \code{sparse_numeric} object.

A numeric vector.
}
\description{
These methods convert between dense numeric vectors and sparse_numeric
vectors.

Converts a dense numeric vector into a \code{sparse_numeric} object by
storing only non-zero entries and their positions.

Converts a \code{sparse_numeric} object into a full dense numeric vector.
}
