% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Calculate mean of a sparse vector}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{...}{Additional arguments (unused)}
}
\value{
The mean value of all elements including zeros
}
\description{
Calculate mean of a sparse vector
}
\examples{
x <- as(c(1, 0, 2, 0, 0), "sparse_numeric")
mean(x)  # Should return 0.6 (sum=3, length=5)

}
