% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{mean}
\alias{mean}
\alias{mean,sparse_numeric-method}
\title{Mean of a sparse_numeric Vector}
\usage{
\S4method{mean}{sparse_numeric}(x)
}
\arguments{
\item{x}{A sparse vector.}
}
\value{
A numeric mean.
}
\description{
Computes the mean including zeros (but without converting to dense form).
}
