% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Calculate Euclidean norm of a sparse vector}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{...}{Additional arguments (unused)}
}
\value{
The Euclidean norm (square root of sum of squared elements)
}
\description{
Calculate Euclidean norm of a sparse vector
}
\examples{
x <- as(c(3, 0, 4, 0), "sparse_numeric")
norm(x)  # Should return 5 (sqrt(3^2 + 4^2))

}
