% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot two sparse vectors for comparison}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{First sparse vector}

\item{y}{Second sparse vector}

\item{...}{Additional plotting parameters}
}
\description{
Plot two sparse vectors for comparison
}
\examples{
\dontrun{
x <- as(c(1, 0, 2, 0, 3), "sparse_numeric")
y <- as(c(0, 2, 0, 1, 0), "sparse_numeric")
plot(x, y)
}

}
