% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{plot}
\alias{plot}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot Two sparse_numeric Vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{First sparse vector.}

\item{y}{Second sparse vector.}
}
\value{
A basic overlap plot.
}
\description{
Produces a simple plot marking positions where both vectors contain
nonzero entries.
}
