% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{show,sparse_numeric-method}
\alias{show,sparse_numeric-method}
\title{Display a sparse vector}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{A sparse_numeric object to display}
}
\description{
Display a sparse vector
}
