% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{show}
\alias{show}
\alias{show,sparse_numeric-method}
\title{Display a sparse_numeric Vector}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{A \code{sparse_numeric} vector.}
}
\value{
Printed output.
}
\description{
Prints the vector in a readable line-by-line format.
}
