% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sort}
\alias{sort}
\alias{sort,sparse_numeric-method}
\title{Sort a sparse_numeric Object}
\usage{
\S4method{sort}{sparse_numeric}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{decreasing}{Whether to sort in decreasing order of position.}

\item{...}{Additional arguments (unused).}
}
\value{
A new sorted \code{sparse_numeric} vector.
}
\description{
Reorders a sparse vector based on the stored positions.
}
