% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse-arithmetic}
\alias{sparse-arithmetic}
\alias{+,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\title{Arithmetic operators for sparse vectors}
\usage{
\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{First sparse vector}

\item{e2}{Second sparse vector}
}
\value{
A sparse vector containing the result of the operation
}
\description{
Arithmetic operators for sparse vectors
}
