% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_add}
\alias{sparse_add}
\title{Generic function for sparse vector addition}
\usage{
sparse_add(x, y, ...)
}
\arguments{
\item{x}{First sparse vector}

\item{y}{Second sparse vector}

\item{...}{Additional arguments}
}
\value{
A sparse vector containing the sum
}
\description{
Generic function for sparse vector addition
}
