% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Compute cross product (dot product) of two sparse vectors}
\usage{
\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{First sparse vector}

\item{y}{Second sparse vector}

\item{...}{Additional arguments (unused)}
}
\value{
A numeric value representing the dot product
}
\description{
Compute cross product (dot product) of two sparse vectors
}
\examples{
x <- as(c(1, 0, 2, 0), "sparse_numeric")
y <- as(c(0, 3, 0, 4), "sparse_numeric")
result <- sparse_crossprod(x, y)
# Result should be 0 (no overlapping non-zero positions)

}
