% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\title{Generic function for sparse vector cross product}
\usage{
sparse_crossprod(x, y, ...)
}
\arguments{
\item{x}{First sparse vector}

\item{y}{Second sparse vector}

\item{...}{Additional arguments}
}
\value{
A numeric value representing the dot product
}
\description{
Generic function for sparse vector cross product
}
