% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Multiply two sparse vectors element-wise}
\usage{
\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{First sparse vector}

\item{y}{Second sparse vector}

\item{...}{Additional arguments (unused)}
}
\value{
A sparse vector containing the element-wise product
}
\description{
Multiply two sparse vectors element-wise
}
\examples{
x <- as(c(1, 0, 2, 0), "sparse_numeric")
y <- as(c(0, 3, 0, 4), "sparse_numeric")
result <- sparse_mult(x, y)
# Result should be all zeros (no overlapping non-zero positions)

}
