% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_mult}
\alias{sparse_mult}
\title{Generic function for sparse vector multiplication}
\usage{
sparse_mult(x, y, ...)
}
\arguments{
\item{x}{First sparse vector}

\item{y}{Second sparse vector}

\item{...}{Additional arguments}
}
\value{
A sparse vector containing the element-wise product
}
\description{
Generic function for sparse vector multiplication
}
