% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_numeric-arithmetic}
\alias{sparse_numeric-arithmetic}
\alias{+,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{Arithmetic Methods for sparse_numeric}
\usage{
\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{First operand (a sparse_numeric object)}

\item{e2}{Second operand (a sparse_numeric object)}
}
\description{
Arithmetic Methods for sparse_numeric
}
\section{Functions}{
\itemize{
\item \code{e1 + e2}: Addition for sparse vectors.

\item \code{e1 - e2}: Subtraction for sparse vectors.

\item \code{e1 * e2}: Multiplication for sparse vectors.

}}
