% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_numeric-validity}
\alias{sparse_numeric-validity}
\title{Validity check for sparse_numeric objects}
\arguments{
\item{object}{A sparse_numeric object to validate}
}
\value{
TRUE if valid, otherwise an error message
}
\description{
Validity check for sparse_numeric objects
}
