% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtract two sparse vectors}
\usage{
\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{First sparse vector (minuend)}

\item{y}{Second sparse vector (subtrahend)}

\item{...}{Additional arguments (unused)}
}
\value{
A sparse vector containing the element-wise difference
}
\description{
Subtract two sparse vectors
}
\examples{
x <- as(c(1, 0, 2, 0), "sparse_numeric")
y <- as(c(0, 3, 0, 4), "sparse_numeric")
result <- sparse_sub(x, y)
# Result should have values 1, -3, 2, -4 at positions 1, 2, 3, 4

}
