% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_sub}
\alias{sparse_sub}
\title{Generic function for sparse vector subtraction}
\usage{
sparse_sub(x, y, ...)
}
\arguments{
\item{x}{First sparse vector (minuend)}

\item{y}{Second sparse vector (subtrahend)}

\item{...}{Additional arguments}
}
\value{
A sparse vector containing the difference
}
\description{
Generic function for sparse vector subtraction
}
