% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_sum}
\alias{sparse_sum}
\alias{sparse_sum,sparse_numeric-method}
\title{Sum of a sparse_numeric Vector}
\usage{
sparse_sum(object)

\S4method{sparse_sum}{sparse_numeric}(object)
}
\arguments{
\item{object}{A sparse vector.}
}
\value{
The numeric sum.
}
\description{
Computes the sum of all nonzero stored values.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_sum(sparse_numeric)}: Sum stored entries.

}}
