% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparsity}
\alias{sparsity}
\alias{sparsity,sparse_numeric-method}
\title{Calculate sparsity percentage of a sparse vector}
\usage{
sparsity(x)

\S4method{sparsity}{sparse_numeric}(x)
}
\arguments{
\item{x}{A sparse_numeric object}
}
\value{
The percentage of non-zero elements (0-100)
}
\description{
Calculate sparsity percentage of a sparse vector
}
\examples{
x <- as(c(1, 0, 2, 0, 0), "sparse_numeric")
sparsity(x)  # Should return 40 (2 non-zero out of 5 elements)

}
