% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize a sparse vector}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{...}{Additional arguments (unused)}
}
\value{
A standardized sparse vector
}
\description{
Centers the vector by subtracting the mean and scales by dividing by the standard deviation.
}
\examples{
x <- as(c(1, 2, 3, 4, 5), "sparse_numeric")
standardized <- standardize(x)
# Result will have mean approximately 0 and standard deviation 1

}
