% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{*,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{Multiplication of Sparse Numeric Vectors (Operator)}
\usage{
\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{A \code{sparse_numeric} object.}

\item{e2}{A \code{sparse_numeric} object of the same length as \code{e1}.}
}
\value{
A \code{sparse_numeric} object representing the element-wise product.
}
\description{
Computes element-wise multiplication of two \code{sparse_numeric} objects using the \code{*} operator.
}
\examples{
x <- new("sparse_numeric", value = c(1,2), pos = c(2L,5L), length = 6L)
y <- new("sparse_numeric", value = c(3,4), pos = c(1L,5L), length = 6L)
x * y
}
