% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Mean of a Sparse Numeric Vector}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Additional arguments.}
}
\value{
A numeric scalar representing the mean.
}
\description{
Computes the mean of a \code{sparse_numeric} vector, including all the zeros.
}
\examples{
x <- new("sparse_numeric", value = c(1, 2), pos = c(2L, 5L), length = 6L)
mean(x)
}
