% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Mean for sparse_numeric
Computes the mean of all entries in a sparse_numeric vector,
including the implicit zeros.}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object.}

\item{...}{Additional arguments (not used).}
}
\value{
Numeric scalar giving the mean.
}
\description{
Mean for sparse_numeric
Computes the mean of all entries in a sparse_numeric vector,
including the implicit zeros.
}
