% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Norm of a Sparse Numeric Vector}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Additional arguments.}
}
\value{
A numeric scalar representing the squared norm.
}
\description{
Computes the squared norm of a \code{sparse_numeric} vector which is the square root of
the sum of the squared individual elements of the vector.
}
\examples{
x <- new("sparse_numeric", value = c(1, 2), pos = c(2L, 5L), length = 6L)
norm(x)
}
