% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Norm of a sparse_numeric Vector}
\usage{
norm(object)

\S4method{norm}{sparse_numeric}(object)
}
\arguments{
\item{object}{A sparse vector.}
}
\value{
A numeric norm.
}
\description{
Computes the Euclidean (L2) norm using only stored values.
}
\section{Methods (by class)}{
\itemize{
\item \code{norm(sparse_numeric)}: Euclidean norm for sparse vectors.

}}
