% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Compute the Euclidean norm of a sparse numeric vector}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric scalar equal to \code{sqrt(sum(x^2))}.
}
\description{
Compute the Euclidean norm of a sparse numeric vector
}
