% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot Two Sparse Numeric Vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{y}{A \code{sparse_numeric} object.}

\item{...}{Additional arguments.}
}
\value{
\code{NULL}. Produces a scatter plot of overlapping non-zero elements.
}
\description{
Plots the overlapping non-zero elements of two \code{sparse_numeric} vectors.
}
\examples{
x <- new("sparse_numeric", value = c(1,2), pos = c(2L,5L), length = 6L)
y <- new("sparse_numeric", value = c(3,4), pos = c(1L,5L), length = 6L)
plot(x, y)
}
