% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{show,sparse_numeric-method}
\alias{show,sparse_numeric-method}
\title{Show a Sparse Numeric Vector}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{A \code{sparse_numeric} object.}
}
\value{
\code{NULL}. Prints information to the console.
}
\description{
Displays a \code{sparse_numeric} object in a readable form, showing its length
and all non-zero elements with their positions.
}
\examples{
x <- new("sparse_numeric", value = c(1,2), pos = c(2L,5L), length = 6L)
x  # Calls show()
}
