% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_add}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\title{Element-wise Addition for sparse_numeric Vectors}
\usage{
sparse_add(x, y)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{One sparse vector.}

\item{y}{Another sparse vector.}
}
\value{
A \code{sparse_numeric} containing the element-wise sum.
}
\description{
Defines a generic and method to add two sparse vectors without
converting them to dense format.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_add(x = sparse_numeric, y = sparse_numeric)}: Add two sparse vectors.

}}
