% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_add}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\alias{+,sparse_numeric,sparse_numeric-method}
\title{Add two sparse numeric vectors}
\usage{
sparse_add(x, y, ...)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y)

\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{y}{A \code{sparse_numeric} object of the same length as \code{x}.}

\item{...}{Additional arguments (not used).}

\item{e1, e2}{\code{sparse_numeric} objects (used by the \code{+} operator).}
}
\value{
A \code{sparse_numeric} object representing \code{x + y}.
}
\description{
Add two sparse numeric vectors
}
