% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Cross Product of Sparse Numeric Vectors}
\usage{
sparse_crossprod(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{y}{A \code{sparse_numeric} object of the same length as \code{x}.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric scalar representing the cross product of \code{x} and \code{y}.
}
\description{
Computes the cross product of two \code{sparse_numeric} vectors.
}
\examples{
x <- new("sparse_numeric", value = c(1,2), pos = c(2L,5L), length = 6L)
y <- new("sparse_numeric", value = c(3,4), pos = c(1L,5L), length = 6L)
sparse_crossprod(x, y)
}
