% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Dot Product of Two sparse_numeric Vectors}
\usage{
sparse_crossprod(x, y)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{A sparse vector.}

\item{y}{A sparse vector.}
}
\value{
A numeric scalar.
}
\description{
Computes the sum of pairwise products for overlapping nonzero positions.
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_crossprod(x = sparse_numeric, y = sparse_numeric)}: Compute sparse dot product.

}}
