% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Sparse cross product}
\usage{
sparse_crossprod(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{y}{A \code{sparse_numeric} object of the same length as \code{x}.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric scalar equal to \code{sum(x * y)}.
}
\description{
Sparse cross product
}
