% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Multiplication of Sparse Numeric Vectors}
\usage{
sparse_mult(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{y}{A \code{sparse_numeric} object of the same length as \code{x}.}

\item{...}{Additional arguments.}
}
\value{
A \code{sparse_numeric} object representing the element-wise product of \code{x} and \code{y}.
}
\description{
Computes element-wise multiplication of two \code{sparse_numeric} vectors.
}
\examples{
x <- new("sparse_numeric", value = c(1,2), pos = c(2L,5L), length = 6L)
y <- new("sparse_numeric", value = c(3,4), pos = c(1L,5L), length = 6L)
sparse_mult(x, y)
}
