% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Element-wise Multiplication for sparse_numeric Vectors}
\usage{
sparse_mult(x, y)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{A \code{sparse_numeric} vector.}

\item{y}{Another \code{sparse_numeric} vector.}
}
\value{
A \code{sparse_numeric} object containing element-wise products.
}
\description{
Element-wise Multiplication for sparse_numeric Vectors
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_mult(x = sparse_numeric, y = sparse_numeric)}: Multiply two sparse vectors.

}}
