% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse Numeric Vector Class}
\description{
An S4 class to represent numeric vectors that are mostly zero.
}
\section{Slots}{

\describe{
\item{\code{value}}{Numeric vector of non-zero values.}

\item{\code{pos}}{Integer vector of positions corresponding to \code{value}.}

\item{\code{length}}{Integer scalar specifying the total length of the vector.}
}}

