% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse Numeric Vector Class}
\value{
A new \code{sparse_numeric} object.
}
\description{
An S4 class used to represent a numeric vector that contains many zeros.
Only the nonzero entries and their locations are stored.
}
\section{Slots}{

\describe{
\item{\code{value}}{A numeric vector containing all nonzero elements.}

\item{\code{pos}}{An integer vector with the positions of nonzero values.}

\item{\code{length}}{An integer giving the full length of the vector (including zeros).}
}}

