% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_sub}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtract Two sparse_numeric Vectors}
\usage{
sparse_sub(x, y)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{First sparse vector.}

\item{y}{Second sparse vector to subtract.}
}
\value{
A sparse vector representing \code{x - y}.
}
\description{
Subtract Two sparse_numeric Vectors
}
\section{Methods (by class)}{
\itemize{
\item \code{sparse_sub(x = sparse_numeric, y = sparse_numeric)}: Subtract two sparse vectors.

}}
