% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardization of a Sparse Numeric Vector}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Additional arguments.}
}
\value{
A \code{sparse_numeric} object with standardized values.
}
\description{
Computes the standardization of a \code{sparse_numeric} vector which is done by taking
taking each element of the vector and subtracting off the vector mean and dividing
by the vector standard deviation.
}
\examples{
x <- new("sparse_numeric", value = c(1, 2), pos = c(2L, 5L), length = 6L)
z <- standardize(x)
as(z, "numeric")
}
