% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize a sparse_numeric Vector}
\usage{
standardize(object)

\S4method{standardize}{sparse_numeric}(object)
}
\arguments{
\item{object}{A sparse vector.}
}
\value{
A standardized sparse vector.
}
\description{
Converts the sparse vector to dense form, centers it, scales it, and then
converts it back into a sparse representation.
}
\section{Methods (by class)}{
\itemize{
\item \code{standardize(sparse_numeric)}: Standardize a sparse vector.

}}
