% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize a sparse numeric vector}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{sparse_numeric} object representing the standardized vector.
}
\description{
Standardize a sparse numeric vector
}
